bug 13469: adding a warning popup in the widget
If the widget is embedded in a non-PCF site, then when the main menu
is clicked, a warning popup will appear that tells them that the
widget is still experimental in nature.  If they agree, they can
continue on to interact with the widget and a 24-hour cookie will
be set that keeps it from bugging them.