Implemented xml cleaner, fixes issues https://www.pivotaltracker.com/story/show/12996607.
Ok, this fix it but is ugly as hell . Neither Dmitriy's nor mine (the diveintopython one) worked without modification).

a) Because our solution requires doing regex matches on strings, we try first to parse it straight with mindom, then we try our pre treatment ( we should not be processing that much broken xml anyways) - lines 36 -> 41

b) because we're now using lxml, we need to adapt to minidom api (at least the smallest of it) so the rest of the code can function as it is. lines 88 -> 96

c) I've taken a quick look at the vidscraper and I see we are using a very different version. Is this a good idea? Are the needs to different for us to track their master?

d) This fixes only blip for now, if we have more broken xml we should think of a better solution